+++ python/GafferSceneTest/__init__.py
@@ -196,6 +196,7 @@ from .CatalogueSelectTest import CatalogueSelectTest
 from .CameraQueryTest import CameraQueryTest
 from .GlobalsSanitiserTest import GlobalsSanitiserTest
 from .ReflectionConstraintTest import ReflectionConstraintTest
+from .CurvesInterpolationTest import CurvesInterpolationTest
 
 from .IECoreScenePreviewTest import *
 from .IECoreGLPreviewTest import *